V 1.01
Implemented the other currencies.
Currencies that is less worth than Chaos works as intended.
Currencies that is more worth than chaos (decimals) have an issue right now, and might have their own algorithms. As it is right now, the value B and S should be reverted to make the currencies with more worth "work".